requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
easyocr==1.7.2
openpyxl==3.1.5
python-multipart==0.0.20
jinja2==3.1.5
pillow==11.1.0
numpy==1.26.4
